refactor: 调整dataSource数据格式，将行布局改成列布局
