+++ src/app/app.component.html
@@ -9,6 +9,5 @@
 <d-calendar-graph text="Today is 20201103"></d-calendar-graph>
 
 <d-calendar-graph [sourceData]="contributions"></d-calendar-graph>
-<d-calendar-graph [sourceData]="contributionsRow"></d-calendar-graph>
 <d-calendar-graph [sourceData]="contributionsColumn"></d-calendar-graph>
 <router-outlet></router-outlet>
