+++ src/app/calendar-graph/calendar-graph.module.ts
@@ -3,14 +3,18 @@ import { CommonModule } from '@angular/common';
 import { TooltipModule } from 'ng-devui';
 
 import { CalendarGraphComponent } from './calendar-graph.component';
-import { ColorCubeModule } from '../shared/components/color-cube/color-cube.module';
+import { ColorCubeModule } from '@shared/components/color-cube/color-cube.module';
+import { ColorPipeModule } from '@shared/pipes/color-pipe/color-pipe.module';
 
 @NgModule({
-  declarations: [CalendarGraphComponent],
+  declarations: [
+    CalendarGraphComponent
+  ],
   imports: [
     CommonModule,
     ColorCubeModule,
-    TooltipModule
+    TooltipModule,
+    ColorPipeModule
   ],
   exports: [
     CalendarGraphComponent
