+++ src/app/shared/config.ts
@@ -1,0 +1,11 @@
+export const CONTRIBUTION_RANGE = [ 0, 3, 6, 9, 10 ];
+
+// 由浅到深
+export const LEGEND_COLOR = [ '#ebedf0', '#c6e48b', '#7bc96f', '#239a3b', '#196127', '#f00' ];
+
+// 圣诞节主题
+export const LEGEND_COLOR_HALLOWEEN = [ '#ebedf0', '#FFEE4A', '#FFC501', '#FE9600', '#03001C' ];
+
+export const DEFAULT_ROW_NUMBER = 7;
+
+export const DEFAULT_DATE_FORMAT = 'YYYY-MM-DD';
