+++ src/app/shared/pipes/color-pipe/color.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { ColorPipe } from './color.pipe';
+
+describe('ColorPipe', () => {
+  it('create an instance', () => {
+    const pipe = new ColorPipe();
+    expect(pipe).toBeTruthy();
+  });
+});
