+++ src/app/shared/pipes/color-pipe/color.pipe.ts
@@ -1,0 +1,13 @@
+import { Pipe, PipeTransform } from '@angular/core';
+import { getColor } from '@shared/util';
+
+@Pipe({
+  name: 'color'
+})
+export class ColorPipe implements PipeTransform {
+
+  transform(value: number): string {
+    return getColor(value);
+  }
+
+}
